deleted data folder (to be included in gitignore). ask functions moved from AccountingBook class to UserInterface class
