--- data/backup/categories_TestUser.json
@@ -1,3 +1,0 @@
-[
-    "TestCategory"
-]
