--- data/backup/categories_asd.json
@@ -1,3 +1,0 @@
-[
-    "asd"
-]
