--- data/backup/categories_ass.json
@@ -1,3 +1,0 @@
-[
-    "ass"
-]
