--- data/backup/categories_diego.json
@@ -1,5 +1,0 @@
-[
-    "supermercado",
-    "transporte",
-    "hogar"
-]
