--- data/categories_andoni.json
@@ -1,4 +1,0 @@
-[
-    "supermercado",
-    "caca"
-]
