--- data/categories_diego.json
@@ -1,3 +1,0 @@
-[
-    "supermercado"
-]
